% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add}
\alias{add}
\title{This funciton adds up two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A \code{numeric} value.}

\item{y}{A \code{numeric} value.}
}
\value{
A \code{numeric} which adds \code{x} and \code{y}.
}
\description{
What the functionkjh is doing
}
\examples{
add(3,4)
add(0,0)
}
\author{
Stephane Guerrier
}
